+++ CHANGELOG.md
@@ -7,7 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 [Unreleased]: https://github.com/skybet/cali/compare/v0.3.0...master
 ## [Unreleased]
 
-None yet
+### Added
+- Actively catching terminal resize
 
 [0.3.0]:      https://github.com/skybet/cali/compare/v0.2.0...v0.3.0
 ## [0.3.0] - 2018-06-04